Remove the TutorialModel
