+++ tests/unit/testaddon.cpp
@@ -5,7 +5,10 @@
 #include "testaddon.h"
 #include "../../src/addons/addon.h"
 #include "../../src/addons/addonguide.h"
+#include "../../src/addons/addontutorial.h"
 #include "../../src/settingsholder.h"
+#include "../../src/qmlengineholder.h"
+#include "../../src/tutorial/tutorial.h"
 #include "helper.h"
 
 void TestAddon::conditions_data() {
@@ -234,6 +237,149 @@ void TestAddon::guide_create() {
   QCOMPARE(guide->property("image").toString(), "foo.png");
 
   delete guide;
+}
+
+void TestAddon::tutorial_create_data() {
+  QTest::addColumn<QString>("id");
+  QTest::addColumn<QJsonObject>("content");
+  QTest::addColumn<bool>("created");
+
+  QTest::addRow("object-without-id") << "" << QJsonObject() << false;
+
+  QJsonObject obj;
+  obj["id"] = "foo";
+  QTest::addRow("invalid-id") << "foo" << obj << false;
+  QTest::addRow("no-image") << "foo" << obj << false;
+
+  obj["image"] = "foo.png";
+  QTest::addRow("no-steps") << "foo" << obj << false;
+
+  QJsonArray steps;
+  obj["steps"] = steps;
+  QTest::addRow("with-steps") << "foo" << obj << false;
+
+  steps.append("");
+  obj["steps"] = steps;
+  QTest::addRow("with-invalid-step") << "foo" << obj << false;
+
+  QJsonObject step;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-without-id") << "foo" << obj << false;
+
+  step["id"] = "s1";
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-without-element") << "foo" << obj << false;
+
+  step["element"] = "wow";
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-without-next") << "foo" << obj << false;
+
+  step["next"] = "wow";
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next") << "foo" << obj << false;
+
+  QJsonObject nextObj;
+
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-1") << "foo" << obj << false;
+
+  nextObj["op"] = "wow";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-2") << "foo" << obj << false;
+
+  nextObj["op"] = "signal";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-3") << "foo" << obj << false;
+
+  nextObj["signal"] = "a";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-4") << "foo" << obj << false;
+
+  nextObj["qml_emitter"] = "a";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-5") << "foo" << obj << true;
+
+  nextObj["vpn_emitter"] = "a";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-6") << "foo" << obj << false;
+
+  nextObj.remove("qml_emitter");
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-7") << "foo" << obj << false;
+
+  nextObj["vpn_emitter"] = "settingsHolder";
+  step["next"] = nextObj;
+  steps.replace(0, step);
+  obj["steps"] = steps;
+  QTest::addRow("with-step-with-invalid-next-8") << "foo" << obj << true;
+
+  obj["conditions"] = QJsonObject();
+  QTest::addRow("with-step-element and conditions") << "foo" << obj << true;
+}
+
+void TestAddon::tutorial_create() {
+  QFETCH(QString, id);
+  QFETCH(QJsonObject, content);
+  QFETCH(bool, created);
+
+  SettingsHolder settingsHolder;
+
+  QJsonObject obj;
+  obj["tutorial"] = content;
+
+  Addon* tutorial = AddonTutorial::create(nullptr, "foo", "bar", "name", obj);
+  QCOMPARE(!!tutorial, created);
+
+  if (!tutorial) {
+    return;
+  }
+
+  Tutorial* tm = Tutorial::instance();
+  QVERIFY(!!tm);
+  QVERIFY(!tm->isPlaying());
+
+  QString tutorialTitleId = tutorial->property("titleId").toString();
+  QCOMPARE(tutorialTitleId, QString("tutorial.%1.title").arg(id));
+
+  QString tutorialSubtitleId = tutorial->property("subtitleId").toString();
+  QCOMPARE(tutorialSubtitleId, QString("tutorial.%1.subtitle").arg(id));
+
+  QString tutorialCompletionMessageId =
+      tutorial->property("completionMessageId").toString();
+  QCOMPARE(tutorialCompletionMessageId,
+           QString("tutorial.%1.completion_message").arg(id));
+
+  QCOMPARE(tutorial->property("image").toString(), "foo.png");
+
+  QmlEngineHolder qml;
+
+  QSignalSpy signalSpy(tm, &Tutorial::playingChanged);
+
+  tm->play(static_cast<AddonTutorial*>(tutorial));
+  QCOMPARE(signalSpy.count(), 1);
+
+  tm->stop();
+  QCOMPARE(signalSpy.count(), 2);
+
+  delete tutorial;
 }
 
 static TestAddon s_testAddon;
